adding docker-compose configuration.
This way, the interactive session can be started with

    docker-compose run tutorial

from the host system's repository folder.

The repository folder will be mounted into the container such that the scripts can be edited from the host system and the results will be accessible on the host system without using a mounted `/tmp` folder.